Made the commit Graph significantly faster by using Git's --date-order argument.
This is not as *accurate* as doing it ourselves, because Git will preemptively emit commits (before it can be proved that they are next in date-order) when using the -n argument.  To compensate, we seek 1000 commits further than requested.  It is highly unlikely that anyone would have run in to the issue in the first place, which makes this a very reasonable compromise of performance and accuracy.